# BatchRun/static/run_swagger.yaml
openapi: 3.0.2
info:
  title: Batch RUN API
  description: API for managing batch runs
  version: 1.0.0
servers:
  - url: http://localhost:5000
paths:
  /run:
    post:
      summary: Start a new run
      description: Start a new run with the provided parameters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
                - cob_date
                - run_group
                - scenario
              properties:
                type:
                  type: string
                  enum:
                    - CCAR
                    - CECL
                    - IFRS
                    - RiskAppetite
                  description: Type of run
                cob_date:
                  type: string
                  format: date
                  example: "2023-03-09"
                  description: COB date
                run_group:
                  type: string
                  example: "Group1"
                  description: Run group
                scenario:
                  type: string
                  example: "Scenario1"
                  description: Scenario
      responses:
        201:
          description: Run started successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id:
                    type: string
                    example: "1234567890"
                    description: ID of the started run
        400:
          description: Invalid request
        500:
          description: Internal server error
  /run/{run_id}/status:
    get:
      summary: Get the status of a run
      description: Get the status of a run with the provided ID
      parameters:
        - in: path
          name: run_id
          schema:
            type: string
          required: true
          description: ID of the run
      responses:
        200:
          description: Run status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "running"
                  progress:
                    type: number
                    example: 0.5
        404:
          description: Run not found
        500:
          description: Internal server error
  /run/{run_id}/kill:

    post:
      summary: Kill a run
      description: Kill a run with the provided ID
      parameters:
        - in: path
          name: run_id
          schema:
            type: string
          required: true
          description: ID of the run
      responses:
        200:
          description: Run killed successfully
        404:
          description: Run not found
        500:
          description: Internal server error
  /runInfo:
    get:
      summary: Get filtered records
      description: To retrieve CECL unadjusted results from a single calculator (e.g. either CNI or CRE)
      operationId: getFilteredRecords
      parameters:
        - name: cob_date
          in: query
          description: Format = yyyy-mm-dd (e.g., 2024-06-30)
          required: false
          schema:
            type: string
            example: "2024-06-30"
        - name: portfolio
          in: query
          description: Portfolio name (e.g., CNI, CRE, BB)
          required: false
          schema:
            type: string
            example: "CNI"
        - name: run_group
          in: query
          description: Run group type (e.g., official, prelim)
          required: false
          schema:
            type: string
            example: "official"
        - name: batch_id
          in: query
          description: Unique batch ID
          required: false
          schema:
            type: string
            example: "BATCH12345"
      responses:
        '200':
          description: Filtered records based on query parameters
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    cob_date:
                      type: string
                      example: "2024-06-30"
                    portfolio:
                      type: string
                      example: "CNI"
                    lob:
                      type: string
                      example: "CB"
                    sub_lob:
                      type: string
                      example: "AWM"
                    used_exposure_amount:
                      type: number
                      example: 5000000
                    unused_exposure_amount:
                      type: number
                      example: 2000000
                    loss_given_default:
                      type: number
                      example: 0.3
                    maturity_date:
                      type: string
                      example: "2029-06-30"
                    batch_id:
                      type: string
                      example: "BATCH12345"
                    run_group:
                      type: string
                      example: "official"
        '400':
          description: Invalid request parameters
        '404':
          description: No records found
components:
  schemas:
    Run:
      type: object
      properties:
        run_id:
          type: string
          example: "1234567890"
          description: ID of the run
        type:
          type: string
          example: "CCAR"
          description: Type of run
        cob_date:
          type: string
          format: date
          example: "2023-03-09"
          description: COB date
        run_group:
          type: string
          example: "Group1"
          description: Run group
        scenario:
          type: string
          example: "Scenario1"
          description: Scenario
        status:
          type: string
          example: "running"
          description: Status of the run
        progress:
          type: number
          example: 0.5
          description: Progress of the run